requirements_dev.txt
tensorflow==2.2.0
pyserial==3.4
imutils==0.5.3
opencv-python==4.2.0.34
pandas==1.0.3
